moves testing util to testing package
